Common request error handler on frontend
* Refactoring: move `getProject`, `getOrganization` to Requests.kt
